requirements.txt
requests
python-dotenv
flask
vt-py